Move filename helper method to base writer
