+++ airbyte-integrations/connectors/destination-s3/src/main/java/io/airbyte/integrations/destination/s3/S3Format.java
@@ -25,6 +25,16 @@
 package io.airbyte.integrations.destination.s3;
 
 public enum S3Format {
-  CSV,
-  PARQUET
+  CSV("csv"),
+  PARQUET("parquet");
+
+  private final String fileExtension;
+
+  S3Format(String fileExtension) {
+    this.fileExtension = fileExtension;
+  }
+
+  public String getFileExtension() {
+    return fileExtension;
+  }
 }
